requirements.txt
streamlit>=1.33,<2.0
pandas>=2.1,<3.0
openpyxl>=3.1,<4.0
python-docx>=0.8.11,<1.0
